Add SIGTERM handlers to other components
