+++ janus_server/src/binary_utils.rs
@@ -13,6 +13,7 @@ use backoff::{future::retry, ExponentialBackoff};
 use base64::STANDARD_NO_PAD;
 use deadpool::managed::TimeoutType;
 use deadpool_postgres::{Manager, Pool, PoolError, Runtime, Timeouts};
+use futures::StreamExt;
 use http::StatusCode;
 use janus_core::time::Clock;
 use ring::aead::{LessSafeKey, UnboundKey, AES_128_GCM};
@@ -321,4 +322,33 @@ async fn health_endpoint_server(address: SocketAddr) {
         .map(|| warp::reply::with_status(warp::reply::reply(), StatusCode::OK));
     let server = warp::serve(filter);
     server.bind(address).await;
+}
+
+/// Register a signal handler for SIGTERM, and return a future that will become ready when a
+/// SIGTERM signal is received.
+pub fn setup_signal_handler() -> Result<impl Future<Output = ()>, std::io::Error> {
+    let mut signal_stream = signal_hook_tokio::Signals::new([signal_hook::consts::SIGTERM])?;
+    let handle = signal_stream.handle();
+    let (sender, receiver) = futures::channel::oneshot::channel();
+    let mut sender = Some(sender);
+    tokio::spawn(async move {
+        while let Some(signal) = signal_stream.next().await {
+            if signal == signal_hook::consts::SIGTERM {
+                if let Some(sender) = sender.take() {
+                    // This may return Err(()) if the receiver has been dropped already. If
+                    // that is the case, the consumer must be shut down already, so we can
+                    // safely ignore the error case.
+                    let _ = sender.send(());
+                    handle.close();
+                    break;
+                }
+            }
+        }
+    });
+    Ok(async move {
+        // The receiver may return Err(Canceled) if the sender has been dropped. By inspection, the
+        // sender always has a message sent across it before it is dropped, and the async task it
+        // is owned by will not terminate before that happens.
+        receiver.await.unwrap_or_default()
+    })
 }
